requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
joblib==1.5.1
scikit-learn==1.7.0
tensorflow-cpu==2.18.0
tf-keras==2.18.0
keras==3.8.0
numpy==1.26.4
opencv-python-headless==4.8.0.74
retina-face==0.0.17
typing-extensions==4.8.0